% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraintMat.R
\name{constraintMat}
\alias{constraintMat}
\title{Generating the costraint matrix}
\usage{
constraintMat(m, maxplace)
}
\arguments{
\item{m}{number of the weights}

\item{maxplace}{the index of the maximum weight}
}
\value{
the constraint matrix
}
\description{
A function to generate the constraint matrix
}
\examples{
\dontrun{
data(windspeed)
tdata<-windspeed$H10
m<-7
maxplace<-maxWeight(m,Fn)
constranintMat(m,maxplace)
}
}
