% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxWeight.R
\name{maxWeight}
\alias{maxWeight}
\title{Finding the index of the maximum weight}
\usage{
maxWeight(m, Fn, lower, upper)
}
\arguments{
\item{m}{number of the weights}

\item{Fn}{emiprical cdf}

\item{lower}{the left boundary}

\item{upper}{the right boundary}
}
\value{
the index of the maximum weight
}
\description{
A function to determine the index of the maximum weight
}
\examples{
\dontrun{
data(windspeed)
tdata<-windspeed$H10
Fn<-ecdf(tdata)
m<-7
maxWeight(m,Fn)
}
}
