\name{SC19030-package}
\alias{SC19030-package}
\alias{SC19030}
\docType{package}
\title{
  A function for unimodal density estimation
}
\description{
  R fucntion for unimodal density estimation _umd_ and other four function needed in _umd are provided in this package.
}
\author{
Xiaolin Bo

Maintainer: Xiaolin Bo <742905761@qq.com>
}
\references{
  Turnbull B C, Ghosh S K. Unimodal density estimation using Bernstein polynomials[J]. Computational Statistics & Data Analysis, 2014, 72: 13-29.
}
\keyword{ package }

