% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solveWeights.R
\name{solveWeights}
\alias{solveWeights}
\title{Finding the weight vector}
\usage{
solveWeights(m, Fn, lower, upper, Dmat, dvec)
}
\arguments{
\item{m}{the number of weights}

\item{Fn}{empirical cdf}

\item{lower}{the left boundary}

\item{upper}{the right boundary}

\item{Dmat}{some constraint matrix for calculating}

\item{dvec}{some constraint vector for calculating}
}
\value{
a vector of weights
}
\description{
A function to solve for the weight vector
}
\examples{
\dontrun{
data(windspeed)
tdata<-windspeed$H10
m<-7
ep <- 3/( 8*length(tdata) )	
Fn<-ecdf(tdata)
ecdf.vec<-Fn(tdata)
L <- diag( n / ( ( ecdf.vec + ep)*(1 + ep - ecdf.vec ) ) )
B <- NULL
for(k in 1:m){
 B <- cbind(B, pbeta(tdata, shape1=k, shape2=m-k+1))
}
Dmat <- t(B) \%*\% L \%*\% B
dvec <- t(B)\%*\% L \%*\% ecdf.vec
solveWeights(m,Fn,Dmat,dvec)
}
}
