% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umd.R
\name{umd}
\alias{umd}
\title{Use Bernstein polynomials to achieve unimodal density estimation using R}
\arguments{
\item{data}{the data to be estimated (numeric)}

\item{bound.type}{the type of bound (logical)}

\item{fix.lower}{the minimum of the estimation region (lnumeric)}

\item{fix.upper}{the maximum of the estimation region (numeric)}

\item{crit}{methods for selecting the number of weights (logical)}

\item{m}{the number of weights (numeric)}
}
\value{
a list of weights, optimal number of weights, standard dendity function, lower and upper value
}
\description{
A density estimation function with unimodal constraint
}
\examples{
\dontrun{
data(windspeed)
h10<-windspeed$H10
umd(h10, bound.type="sd", crit="CN")
}
}
